rewrite tests for cobra
