+++ .goreleaser.yaml
@@ -21,91 +21,6 @@ before:
     - go generate ./...
 
 builds:
-  - id: "datasetIngestor"
-    flags:
-      - -trimpath
-    ldflags:
-      - -s -w -X main.VERSION={{.Version}} # This will set the VERSION variable in the binary to the github tag
-    env:
-      - CGO_ENABLED=0
-    dir: ./cmd/datasetIngestor/
-    goos:
-      - linux
-      - windows
-      - darwin
-    goarch:
-      - amd64
-    main: .
-    binary: datasetIngestor
-
-  - id: "datasetArchiver"
-    flags:
-      - -trimpath
-    ldflags:
-      - -s -w -X main.VERSION={{.Version}}
-    env:
-      - CGO_ENABLED=0
-    dir: ./cmd/datasetArchiver/
-    goos:
-      - linux
-      - windows
-      - darwin
-    goarch:
-      - amd64
-    main: .
-    binary: datasetArchiver
-
-  - id: "datasetRetriever"
-    flags:
-      - -trimpath
-    ldflags:
-      - -s -w -X main.VERSION={{.Version}}
-    env:
-      - CGO_ENABLED=0
-    dir: ./cmd/datasetRetriever/
-    goos:
-      - linux
-      - windows
-      - darwin
-    goarch:
-      - amd64
-    main: .
-    binary: datasetRetriever
-
-  - id: "datasetCleaner"
-    flags:
-      - -trimpath
-    ldflags:
-      - -s -w -X main.VERSION={{.Version}}
-    env:
-      - CGO_ENABLED=0
-    dir: ./cmd/datasetCleaner/
-    goos:
-      - linux
-      - windows
-      - darwin
-    goarch:
-      - amd64
-    main: .
-    binary: datasetCleaner
-
-  - id: "datasetGetProposal"
-    flags:
-      - -trimpath
-    ldflags:
-      - -s -w -X main.VERSION={{.Version}}
-    env:
-      - CGO_ENABLED=0
-    dir: ./cmd/datasetGetProposal/
-    goos:
-      - linux
-      - windows
-      - darwin
-    goarch:
-      - amd64
-    main: .
-    binary: datasetGetProposal
-  
   - id: "scicat-cli"
     flags:
       - -trimpath
